Add files via upload
This website, https://www.profesijuklasifikatorius.lt/lpk-paieska, was mostly used to manually label job descriptions in accordance to the Lithuanian version of the ISCO classificator in order to complete this assignment as precisely as possible. Additionally, the CSV file that is attached is a static version of the previously mentioned website (dynamic version of classificator).